fix: add ZELLIJ_AUTO_START

+++ functions/zellij-auto-start.fish
@@ -5,9 +5,12 @@ function zellij-auto-start
     set current_number (zellij list-sessions | rg "automatic-(\w+)" -or '$1' | sort -r | head -n 1)
     test -z "$current_number" && set current_number 0
     set new_number (math "$current_number" + 1)
-    zellij -s "automatic-$new_number"
 
-    if $ZELLIJ_AUTO_EXIT
-        kill $fish_pid
+    if $ZELLIJ_AUTO_START
+        zellij -s "automatic-$new_number"
+
+        if $ZELLIJ_AUTO_EXIT
+            kill $fish_pid
+        end
     end
 end
